requirements.txt
fastapi
uvicorn
python-multipart
requests
torf
pandas_profiling
python-dotenv
ipython
ipywidgets